+++ ExchangeCalculateApp/ExchangeCalculateApp/Controllers/ViewController.swift
@@ -86,7 +86,7 @@ final class ViewController: UIViewController {
                 await MainActor.run {
                     fetchData(rates: result.rates)
                 }
-            } catch(let error) {
+            } catch {
                 let alert: UIAlertController = .initErrorAlert(title: "오류", message: "데이터를 불러올 수 없습니다.")
                 await MainActor.run {
                     self.present(alert, animated: true)
@@ -194,48 +194,4 @@ extension ViewController: UITableViewDelegate {
 }
 
 
-// MARK: NetworkManager
-struct Response: Codable {
-    let result: String
-    let base: String
-    let rates: [String: Double]
-    
-    enum CodingKeys: String, CodingKey {
-        case result
-        case base = "base_code"
-        case rates
-    }
-}
-
-
-enum NetworkError: Error {
-    case invalidURL
-    case decodingError
-    case serverError(code: Int)
-    case transportError
-}
-
-enum ServerURL {
-    static let string = "https://open.er-api.com/v6/latest/USD"
-}
-
-final class NetworkManager {
-    func fetch<T: Codable>(type: T.Type, for url: String) async throws ->  T {
-        guard let url = URL(string: url) else {
-            throw(NetworkError.invalidURL)
-        }
-        let request = Session.default.request(url)
-        
-        let response = await request.serializingDecodable(T.self).response
-        
-        guard response.response?.statusCode == 200 else {
-            throw NetworkError.serverError(code: response.response?.statusCode ?? 0)
-        }
-        
-        guard let data = response.value else {
-            throw(NetworkError.decodingError)
-        }
-        return data
-    }
-}
